[modify]bug DataError: integer out of range

+++ helloworld/models.py
@@ -4,4 +4,4 @@ from django.db import models
 
 # Create your models here.
 class Account(models.Model):
-    mobile_phone_num = models.IntegerField(unique=True)
+    mobile_phone_num = models.BigIntegerField(unique=True)
